update julia3d: norminette 보완.
