+++ Makefile
@@ -4,16 +4,13 @@ CFLAGS=-Wall -Wextra -Werror -I $(HEADER_DIR)
 RM=rm -fr
 MLX_DIR=./srcs/mlx
 MLX_FLAGS=-L$(MLX_DIR) -lmlx -framework OpenGL -framework AppKit
-SRCS_DIR=./srcs/
-SRCS=$(SRCS_DIR)main.c $(SRCS_DIR)main_utils.c $(SRCS_DIR)main_utils_parse.c $(SRCS_DIR)ui_init_screen.c $(SRCS_DIR)ui_key.c $(SRCS_DIR)ui_mouse.c $(SRCS_DIR)ui_button.c $(SRCS_DIR)ui_button_execute.c $(SRCS_DIR)ui_menu.c $(SRCS_DIR)ui_theme.c $(SRCS_DIR)fractal_utils.c $(SRCS_DIR)fractal_utils_recurrence.c $(SRCS_DIR)fractal_utils_execute.c $(SRCS_DIR)logistic_utils.c $(SRCS_DIR)mandelbrot.c $(SRCS_DIR)julia.c $(SRCS_DIR)burningship.c $(SRCS_DIR)tricorn.c $(SRCS_DIR)logistic_map.c
-OBJS=$(SRCS:.c=.o)
 
 MAN_HEADER_DIR=./srcs/mandatory/headers/
 MAN_HEADERS=$(MAN_HEADER_DIR)main.h $(MAN_HEADER_DIR)ui.h $(MAN_HEADER_DIR)fractal.h
 MAN_SRCS_DIR=./srcs/mandatory/
 MAN_MAINS=$(MAN_SRCS_DIR)main/main.c $(MAN_SRCS_DIR)main/main_utils.c $(MAN_SRCS_DIR)main/main_utils_parse.c 
 MAN_UIS=$(MAN_SRCS_DIR)ui/ui_init_screen.c $(MAN_SRCS_DIR)ui/ui_key.c $(MAN_SRCS_DIR)ui/ui_key_utils.c $(MAN_SRCS_DIR)ui/ui_mouse.c $(MAN_SRCS_DIR)ui/ui_button.c $(MAN_SRCS_DIR)ui/ui_button_execute.c $(MAN_SRCS_DIR)ui/ui_menu.c $(MAN_SRCS_DIR)ui/ui_theme.c $(MAN_SRCS_DIR)ui/ui_menu_utils.c
-MAN_FRACTALS=$(MAN_SRCS_DIR)fractal/fractal_utils.c $(MAN_SRCS_DIR)fractal/fractal_utils_recurrence.c $(MAN_SRCS_DIR)fractal/fractal_utils_execute.c $(MAN_SRCS_DIR)fractal/logistic_utils.c $(MAN_SRCS_DIR)fractal/mandelbrot.c $(MAN_SRCS_DIR)fractal/julia.c $(MAN_SRCS_DIR)fractal/burningship.c $(MAN_SRCS_DIR)fractal/tricorn.c $(MAN_SRCS_DIR)fractal/logistic_map.c $(MAN_SRCS_DIR)fractal/julia3d.c
+MAN_FRACTALS=$(MAN_SRCS_DIR)fractal/fractal_utils.c $(MAN_SRCS_DIR)fractal/fractal_utils_recurrence.c $(MAN_SRCS_DIR)fractal/fractal_utils_execute.c $(MAN_SRCS_DIR)fractal/logistic_utils.c $(MAN_SRCS_DIR)fractal/mandelbrot.c $(MAN_SRCS_DIR)fractal/julia.c $(MAN_SRCS_DIR)fractal/burningship.c $(MAN_SRCS_DIR)fractal/tricorn.c $(MAN_SRCS_DIR)fractal/logistic_map.c $(MAN_SRCS_DIR)fractal/julia3d.c $(MAN_SRCS_DIR)fractal/julia3d_utils.c $(MAN_SRCS_DIR)fractal/julia3d_utils_matrix.c
 MAN_SRCS=$(MAN_MAINS) $(MAN_UIS) $(MAN_FRACTALS)
 MAN_OBJS=$(MAN_SRCS:.c=.o)
