+++ README.md
@@ -15,10 +15,6 @@ Copy and paste the script below to your bash.
 ```bash
 git clone https://github.com/leeluna0476/fract-ol.git fract-ol
 cd fract-ol
-<<<<<<< HEAD
-git checkout personal
-=======
->>>>>>> master
 make
 ./fractol [fractal] [c.real] [c.imag]
 ```
