+++ srcs/mandatory/fractal/fractal_utils.c
@@ -6,7 +6,7 @@
 /*   By: seojilee <seojilee@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/04 08:55:26 by seojilee          #+#    #+#             */
-/*   Updated: 2023/12/07 16:20:11 by seojilee         ###   ########.fr       */
+/*   Updated: 2023/12/12 15:05:44 by seojilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
