+++ srcs/mandatory/headers/fractal.h
@@ -6,7 +6,7 @@
 /*   By: seojilee <seojilee@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 19:42:45 by seojilee          #+#    #+#             */
-/*   Updated: 2023/12/11 20:06:49 by seojilee         ###   ########.fr       */
+/*   Updated: 2023/12/12 15:05:49 by seojilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
