+++ srcs/mandatory/headers/ui.h
@@ -6,7 +6,7 @@
 /*   By: seojilee <seojilee@student.42seoul.>       +# +  +:+       +# +      */
 /*                                                +# +# +# +# +# +   +# +     */
 /*   Created: 2023/12/02 12:59:13 by seojilee          # +#     # +#          */
-/*   Updated: 2023/12/12 12:21:51 by seojilee         ###   ########.fr       */
+/*   Updated: 2023/12/12 15:05:41 by seojilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
